requirements.txt
accelerate==1.2.1
acres==0.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-generator==1.10
async-timeout==5.0.1
attrs==24.3.0
bcrypt==4.3.0
beautifulsoup4==4.11.1
bitsandbytes==0.45.1
blinker==1.8.1
blis==1.1.0
browser-cookie3==0.19.1
bs4==0.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.1.1
cryptography==44.0.3
cycler==0.12.1
cymem==2.0.10
datasets==3.2.0
deepspeed==0.16.3
dill==0.3.8
discord.py==2.4.0
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
ecdsa==0.19.1
einops==0.8.0
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.16.1
fitz==0.0.1.dev2
Flask==3.0.3
Flask-MySQLdb==2.0.0
fonttools==4.43.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
h2==4.1.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.4
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.7.1
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.1
lz4==4.3.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
mdurl==0.1.2
motor==3.7.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
mysql-connector-python==8.4.0
mysqlclient==2.2.4
narwhals==1.18.4
networkx==3.3
nibabel==5.3.2
ninja==1.11.1.3
nipype==1.9.2
nltk==3.8.1
numpy==1.23.4
nvidia-ml-py==12.570.86
openai==0.28.0
openpyxl==3.0.10
outcome==1.2.0
packaging==23.2
pandas==1.5.0
passlib==1.7.4
pathlib==1.0.1
peft==0.14.0
Pillow==10.1.0
preshed==3.0.9
propcache==0.2.0
protobuf==5.29.1
prov==2.0.1
psutil==6.1.1
psycopg2==2.9.10
puremagic==1.28
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.4
PyGetWindow==0.0.9
Pygments==2.18.0
pymongo==4.12.1
PyMsgBox==1.0.9
PyMuPDF==1.25.1
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-gemini-api==2.4.12
python-jose==3.4.0
pytweening==1.0.4
pytz==2022.4
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.4.16
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.5.0
sentence-transformers==3.4.1
shellingham==1.5.4
simplejson==3.19.3
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
starlette==0.46.2
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.3
threadpoolctl==3.4.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traits==6.4.3
transformers==4.48.0
trio==0.22.0
trio-websocket==0.9.2
typer==0.15.1
typing_extensions==4.12.2
urllib3==1.26.12
uvicorn==0.34.2
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webdriver-manager==4.0.2
Werkzeug==3.0.2
win32-setctime==1.1.0
wrapt==1.17.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.0
